add session detection to not rerender the welcome screen
